requirements.txt
asgiref==3.3.1
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
django==3.1.6
djangorestframework==3.12.2
idna==2.10
psycopg2==2.8.6
pycparser==2.20
pyopenssl==20.0.1
pysocks==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
coreapi==2.3.3
coreapi-cli==1.0.9
